v0.1
Worked on this over a couple days, it's a bot to help me beat all my friends in squabble :). Ran the algorithm across the entire wordle word base that I found from another GitHub (). On the wordle word bank it takes ~3.6 guesses to find the solution. On the big word bank which is roughly 10x the size it goes up to ~4.7 guesses. I believe a few improvements on my word ranking algorithms would help me create better guesses after we slim down the list of available words. I believe it is possible to get the wordle word bank down to ~3 guesses for solution. However this would take an entire overhaul of my selection algorithm so I'm not certain ill do it. anyways, feel free to use it to beat your friends in squabble :)